move Set TechPreviewNoUpgrade on the target cluster  to host discovery step
